Factor euler angle code into submodule.
